src/requirements.txt
fastapi==0.53.2
uvicorn==0.11.3
databases[postgresql]==0.2.6
paho-mqtt==1.5.0
httpx==0.12.1

# dev
pytest==5.3.2
requests==2.23.0